fix: safe integer casts, empty credential rejection, task result type
- computer_use: get_optional_coords uses i32::try_from instead of as-cast
- tools/tasks: u64→usize via try_from for offset/limit params
- tools/file_read: u64→usize via try_from for offset/limit params
- rate_limiter: compare failures.len() in usize space (not u32)
- gateway/ws_server: reject empty token/password strings in extract_credentials
- KnowledgeTab: AgentTaskState progress is number|null, result is unknown|null
  to match Rust Option<u8> and Option<serde_json::Value>; render result via
  JSON.stringify when not a plain string

+++ src-tauri/src/rate_limiter.rs
@@ -104,7 +104,7 @@ impl RateLimiter {
         record.failures.push(now);
 
         // Check if threshold exceeded
-        if record.failures.len() as u32 >= self.config.max_attempts {
+        if record.failures.len() >= self.config.max_attempts as usize {
             let lockout_duration = std::time::Duration::from_secs(self.config.lockout_secs);
             record.locked_until = Some(now + lockout_duration);
             warn!(
